Move log types to the eth-rpc
