+++ substrate/frame/revive/rpc/src/types/log.rs
@@ -1,0 +1,62 @@
+// This file is part of Substrate.
+
+// Copyright (C) Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::*;
+use serde::{Deserialize, Serialize};
+
+/// Filter results
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(untagged)]
+pub enum FilterResults {
+	/// new block or transaction hashes
+	Hashes(Vec<H256>),
+	/// new logs
+	Logs(Vec<Log>),
+}
+
+impl Default for FilterResults {
+	fn default() -> Self {
+		FilterResults::Hashes(Default::default())
+	}
+}
+
+/// log
+#[derive(Debug, Default, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct Log {
+	/// address
+	pub address: Address,
+	/// block hash
+	pub block_hash: H256,
+	/// block number
+	pub block_number: U256,
+	/// data
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub data: Option<Bytes>,
+	/// log index
+	pub log_index: U256,
+	/// removed
+	#[serde(default)]
+	pub removed: bool,
+	/// topics
+	#[serde(default)]
+	pub topics: Vec<H256>,
+	/// transaction hash
+	pub transaction_hash: H256,
+	/// transaction index
+	pub transaction_index: U256,
+}
